pyproject.toml
[project]
name = "freesound-audio-tagging-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
# Enable important rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations (type hints)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "NPY",    # NumPy-specific rules
    "PD",     # pandas-vet
    "PT",     # flake8-pytest-style
    "RUF",    # Ruff-specific rules
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PERF",   # Perflint (performance)
]

ignore = [
    "ANN401",  # Dynamically typed expressions (Any) are disallowed
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN", "NPY002", "E501"]  # Relaxed rules for tests
"src/devel.py" = ["ANN"]   # Development/exploration scripts can be more relaxed
"src/features/signal_tools.py" = ["PT028"]  # Test utility functions (not pytest tests)
"src/scripts/devel.py" = ["PT028", "E501"]  # Development script - relaxed
"src/scripts/explore.py" = ["E501"]  # Exploration script - relaxed line length
"src/scripts/train.py" = ["NPY002"]  # Global seed setting for reproducibility
"src/training/metrics.py" = ["NPY002"]  # Legacy numpy in __main__ test block is OK
"**/losses.py" = ["N812"]  # Allow F for torch.nn.functional (PyTorch convention)
"**/spectrogram_optimized.py" = ["N812"]  # Allow F for torch.nn.functional (PyTorch convention)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
